chore: convert lib to js
